Link entry to checker in tools/workflows
