+++ src/app/myworkflows/myworkflows.component.html
@@ -1,84 +1,3 @@
-<!--
-  ~    Copyright 2017 OICR
-  ~
-  ~    Licensed under the Apache License, Version 2.0 (the "License");
-  ~    you may not use this file except in compliance with the License.
-  ~    You may obtain a copy of the License at
-  ~
-  ~        http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~    Unless required by applicable law or agreed to in writing, software
-  ~    distributed under the License is distributed on an "AS IS" BASIS,
-  ~    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~    See the License for the specific language governing permissions and
-  ~    limitations under the License.
-  -->
-
-<app-header>
-  <img src="../assets/images/dockstore/dockstore-workflows-green.png"> My Workflows
-  <div class="pull-right refresh-all-holder">
-    <button type="button" class="refresh-all-button btn btn-primary" (click)="refreshAllWorkflows()" [disabled]="refreshMessage !== null">
-      Refresh All Workflows
-    </button>
-  </div>
-</app-header>
-<div class="container padding-top">
-  <app-alert></app-alert>
-  <div class="row" *ngIf="orgWorkflows?.length == 0">
-    <div class="col-md-12">
-      <div class="alert alert-info" role="alert">
-        <p>
-          <span class="glyphicon glyphicon-info-sign"></span>
-          There are currently no workflows registered under this account, to add your first workflow, do one of the following:
-        </p>
-        <ul>
-          <li>
-            Press the 'Refresh All Workflows' button to create workflow stubs for all of your source control repositories.  Browse to the workflow stubs and refresh/publish those that correspond to real workflows.
-          </li>
-          <li>
-            Press the 'Register Workflow' button to manually add a workflow.
-          </li>
-        </ul>
-        <div *ngIf="!hasGitHubToken">
-          <a class="btn btn-block btn-social btn-github" (click)="link()">
-            <span class="fa fa-github"></span> You do not have a linked GitHub token, click here to link a GitHub token
-          </a>
-        </div>
-      </div>
-    </div>
-  </div>
-  <!--Tool Content Page-->
-  <div class="row">
-    <div class="col-md-3 containers-rsb">
-      <accordion [closeOthers]="oneAtATime">
-        <accordion-group *ngFor="let orgObj of orgWorkflows"[heading]="orgObj?.sourceControl + '/' + orgObj?.organization" panelClass="containers-accordion" [isOpen]="orgObj?.isFirstOpen">
-          <div *ngFor="let workflowObj of orgObj?.workflows">
-            <div class="panel-container-label" [ngClass]="{selected: workflow.id === workflowObj.id}">
-              <div class="container-name-oflw pull-left" title="{{workflowObj?.repository + (workflowObj?.workflowName ? ' (' + workflowObj?.workflowName + ')' : '')}}">
-                <a style="cursor: pointer;" (click)="selectWorkflow(workflowObj)">
-                  {{workflowObj?.repository + (workflowObj?.workflowName ? ' (' + workflowObj?.workflowName + ')' : '')}}
-                </a>
-              </div>
-              <div class="pull-right">
-                <span class="label" [ngClass]="workflowObj?.is_published ? 'label-primary' : 'label-warning'">
-                  {{ workflowObj?.is_published ? 'PUB' : 'UNPUB'}}
-                </span>
-              </div>
-            </div>
-          </div>
-          <hr class="accordion-hr">
-          <app-refresh-workflow-organization class="pull-right" [organization]="orgObj?.organization"></app-refresh-workflow-organization>
-          <button type="button" class="glyphicon glyphicon-plus btn btn-primary btn-xs pull-right add-container-ns" (click)="setModalGitURL(orgObj?.organization + '/new_workflow'); showModal()">
-          </button>
-        </accordion-group>
-      </accordion>
-      <div>
-        <button type="button" class="btn btn-primary btn-sm pull-right" (click)="showModal()" [disabled]="refreshMessage !== null">Register Workflow</button>
-      </div>
-      <app-register-workflow-modal></app-register-workflow-modal>
-    </div>
-    <div *ngIf="workflow" class="col-md-9 containers-rsb">
-      <app-workflow [isWorkflowPublic]="false"></app-workflow>
-    </div>
-  </div>
+<div>
+    <router-outlet></router-outlet>
 </div>
