+++ src/app/myworkflows/myworkflows.routing.ts
@@ -1,3 +1,4 @@
+import { MyWorkflowComponent } from './my-workflow/my-workflow.component';
 /**
  *    Copyright 2017 OICR
  *
@@ -20,8 +21,8 @@ import { MyWorkflowsComponent } from './myworkflows.component';
 const MYTOOLS_ROUTES: Routes = [
   {
     path: '', component: MyWorkflowsComponent, children: [
-      { path: '', component: MyWorkflowsComponent},
-      { path: '**', component: MyWorkflowsComponent}
+      { path: '', component: MyWorkflowComponent},
+      { path: '**', component: MyWorkflowComponent}
     ]
   }
 ];
